+++ ambient_toolbox/mixins/models.py
@@ -12,6 +12,7 @@ class PermissionModelMixin:
     class Meta:
         # No database table creation or deletion operations will be performed for this model.
         managed = False
+        default_permissions = []
 
 
 class SaveWithoutSignalsMixin: