+++ ambient_toolbox/tests/mixins.py
@@ -88,7 +88,7 @@ class RequestProviderMixin:
         request = factory.get(url)
 
         # Set user object if it is of a valid type
-        if user is None or isinstance(user, AbstractBaseUser) or isinstance(user, AnonymousUser):
+        if isinstance(user, AbstractBaseUser | AnonymousUser):
             request.user = user
         else:
             raise ValueError(_('Please pass a user object to RequestProviderMixin.'))